Refactor Videotube dashboard helpers
